Modified sundance-compatible build scripts to use Heidi's new expat configuration options.
Fixed several minor warnings in TSFExtended.

+++ packages/TSFExtended/src/Operators/TSFComposedOperator.hpp
@@ -24,7 +24,7 @@
 // Questions? Contact Michael A. Heroux (maherou@sandia.gov) 
 // 
 // **********************************************************************/
- /* @HEADER@ */
+/* @HEADER@ */
 
 #ifndef TSFCOMPOSEDOPERATOR_HPP
 #define TSFCOMPOSEDOPERATOR_HPP
@@ -49,7 +49,7 @@ namespace TSFExtended
   template <class Scalar> 
   class ComposedOperator : public OpDescribableByTypeID<Scalar>,
                            public Handleable<TSFCore::LinearOp<Scalar> >,
-			   public RowAccessibleOp<Scalar>
+                           public RowAccessibleOp<Scalar>
   {
   public:
     GET_RCP(TSFCore::LinearOp<Scalar>);
@@ -60,13 +60,13 @@ namespace TSFExtended
     ComposedOperator(const LinearOperator<Scalar>& left, 
                      const LinearOperator<Scalar>& right)
       : left_(left), 
-	right_(right),
-	domain_(right.domain()),
-	range_(left.range())
+        right_(right),
+        domain_(right.domain()),
+        range_(left.range())
     {
       TEST_FOR_EXCEPTION(left_.domain() != right_.range(), runtime_error,
-			 "Operators used to construct CompsedOperator are "
-			 << " not compatible");
+                         "Operators used to construct CompsedOperator are "
+                         << " not compatible");
     }
 
     /** Virtual dtor */
@@ -113,8 +113,8 @@ namespace TSFExtended
 
     /** Return the kth row  */
     void getRow(const int& row, 
-		Teuchos::Array<int>& indices, 
-		Teuchos::Array<Scalar>& values) const
+                Teuchos::Array<int>& indices, 
+                Teuchos::Array<Scalar>& values) const
     {
       /*  The algorithm is as follows: get the "row" row of left_ and
           then compute the linear combination of rows of right_
@@ -142,13 +142,13 @@ namespace TSFExtended
           linear combination.  */
       right_.getRow(Lind[0], indices, values);
       sort(indices, values);
-      for (int i = 0; i < values.size(); i++)
+      for (unsigned int i = 0; i < values.size(); i++)
         {
           values[i] = values[i] * Lval[0];
         }
 
       /*  Main loop to do the linear combination. */
-      for (int j = 1; j < Lind.size(); j++)
+      for (unsigned int j = 1; j < Lind.size(); j++)
         {
           Rind.resize(0);
           Rval.resize(0);
@@ -170,23 +170,23 @@ namespace TSFExtended
 
 
     void indAxpy(Teuchos::Array<int>& indices, 
-		 Teuchos::Array<double>& values, double& a,
-		 Teuchos::Array<int>& Rind, 
-		 Teuchos::Array<double>& Rval) const
+                 Teuchos::Array<double>& values, double& a,
+                 Teuchos::Array<int>& Rind, 
+                 Teuchos::Array<double>& Rval) const
     {
       /*  Performs an "indexed axpy."  That is it computes y = ax + y
-	  where y is defined by indices and values and x is defined by
-	  Rind and Rval.  To do this, it copies indices and values into
-	  Tind and Tval and rebuilds indices and values. */
+          where y is defined by indices and values and x is defined by
+          Rind and Rval.  To do this, it copies indices and values into
+          Tind and Tval and rebuilds indices and values. */
 
       sort(indices, values);
       Teuchos::Array<int> Tind;
       Teuchos::Array<double> Tval;
-      for (int i = 0; i < indices.size(); i++)
-	{
-	  Tind.append(indices[i]);
-	  Tval.append(values[i]);
-	}
+      for (unsigned int i = 0; i < indices.size(); i++)
+        {
+          Tind.append(indices[i]);
+          Tval.append(values[i]);
+        }
       indices.resize(0);
       values.resize(0);
       sort(Rind, Rval);
@@ -199,7 +199,7 @@ namespace TSFExtended
       double vR = 0.0;
 
       /* getNext gets the next pair from the specified arrays.  It
-	 returns R (or T) = -1 if there are no more elements. */
+         returns R (or T) = -1 if there are no more elements. */
 
       getNext(kR, R, vR, Rind, Rval);
       getNext(kT, T, vT, Tind, Tval);
@@ -207,51 +207,51 @@ namespace TSFExtended
       /*  Main loop to rebuild indices and values. */
       int large = Rind.size() + Tind.size();
       for (int i = 0; i < large; i++)
-	{
-	  if (R == -1 && T == -1)
-	    {
-	      break;
-	    }
-	  else if (T == -1 || (R != -1 && T > R))
-	    {
-	      indices.append(R);
-	      values.append(vR * a);
-	      getNext(kR, R, vR, Rind, Rval);
-	    }
-	  else if (R == -1 || T < R)
-	    {
-	      indices.append(T);
-	      values.append(vT);
-	      getNext(kT, T, vT, Tind, Tval);
-	    }
-	  else if (T == R)
-	    {
-	      indices.append(T);
-	      values.append(vT + a * vR);
-	      getNext(kR, R, vR, Rind, Rval);
-	      getNext(kT, T, vT, Tind, Tval);
-	    }
-	}
+        {
+          if (R == -1 && T == -1)
+            {
+              break;
+            }
+          else if (T == -1 || (R != -1 && T > R))
+            {
+              indices.append(R);
+              values.append(vR * a);
+              getNext(kR, R, vR, Rind, Rval);
+            }
+          else if (R == -1 || T < R)
+            {
+              indices.append(T);
+              values.append(vT);
+              getNext(kT, T, vT, Tind, Tval);
+            }
+          else if (T == R)
+            {
+              indices.append(T);
+              values.append(vT + a * vR);
+              getNext(kR, R, vR, Rind, Rval);
+              getNext(kT, T, vT, Tind, Tval);
+            }
+        }
     }
 
 
 
 
     void getNext(int& k, int& LR, double& vLR,
-		 Teuchos::Array<int>& ind,
-		 Teuchos::Array<double>& val) const
+                 Teuchos::Array<int>& ind,
+                 Teuchos::Array<double>& val) const
     {
       /* gets next pair from the given arrays.  Returns -1 in LR if no
-	 more elements.  */
-      if (k >= ind.size())
-	{
-	  LR = -1;
-	}
+         more elements.  */
+      if ((unsigned int) k >= ind.size())
+        {
+          LR = -1;
+        }
       else
-	{
-	  LR = ind[k];
-	  vLR = val[k];
-	}
+        {
+          LR = ind[k];
+          vLR = val[k];
+        }
       k++;
       return;
     }
@@ -261,29 +261,29 @@ namespace TSFExtended
     void sort(Teuchos::Array<int>& ind, Teuchos::Array<double>& val) const
     {
       /* Standard bubble sort from bottom, since that's where any changes
-	 are made.  */
+         are made.  */
       bool flip = false;
       int len = ind.size();
       for (int j = 0; j < len;  j++)
-	{
-	  flip = false;
-	  for (int i = len-1; i > j; i--)
-	    {
-	      if (ind[i] < ind[i-1])
-		{
-		  flip = true;
-		  int t = ind[i-1];
-		  ind[i-1] = ind[i];
-		  ind[i] = t;
+        {
+          flip = false;
+          for (int i = len-1; i > j; i--)
+            {
+              if (ind[i] < ind[i-1])
+                {
+                  flip = true;
+                  int t = ind[i-1];
+                  ind[i-1] = ind[i];
+                  ind[i] = t;
 
-		  double v = val[i-1];
-		  val[i-1] = val[i];
-		  val[i] = v;
-		}
-	    }
+                  double v = val[i-1];
+                  val[i-1] = val[i];
+                  val[i] = v;
+                }
+            }
 
-	  if (!flip) return;
-	}
+          if (!flip) return;
+        }
     }
